requirements.txt
streamlit
psycopg2
google-genai
python-dotenv